app/requirements.txt
psycopg2-binary
mnemonic
requests
hedera-sdk-python
ed25519
cryptography
bip32utils

# High-performance optimizations
toml>=0.10.2             # System specs configuration
psutil>=5.9.0            # System monitoring and performance metrics
numpy>=1.24.0            # Numerical operations for GPU acceleration
cachetools>=5.3.0        # Advanced caching mechanisms
aiohttp>=3.8.0           # Async HTTP for concurrent API calls
aiofiles>=23.1.0         # Async file operations

# GPU acceleration (optional but recommended for AMD RX 6600)
pyopencl>=2024.1         # OpenCL support for GPU acceleration

# Optional GPU monitoring
pynvml>=11.0.0           # NVIDIA GPU monitoring (if NVIDIA GPU present)
